refactor(data-development): move stable roles into business packages
