+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/controller/v1/DevelopmentDatasetNodeController.java
@@ -5,8 +5,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import io.yak.framework.common.Result;
 import io.yak.ops.business.dataset.DevelopmentDatasetFacade.FieldDraft;
 import io.yak.ops.business.dataset.DevelopmentDatasetFacade.PreviewResult;
-import io.yak.ops.business.development.service.DevelopmentDatasetNodeService;
-import io.yak.ops.business.development.service.DevelopmentDatasetNodeService.DatasetNodeContext;
+import io.yak.ops.business.development.dataset.DevelopmentDatasetNodeService;
+import io.yak.ops.business.development.dataset.DevelopmentDatasetNodeService.DatasetNodeContext;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
