+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/controller/v1/DevelopmentDirectoryController.java
@@ -5,8 +5,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import io.yak.framework.common.Result;
 import io.yak.ops.business.development.api.DevelopmentDirectoryApi.CreateRequest;
 import io.yak.ops.business.development.api.DevelopmentDirectoryApi.RenameRequest;
+import io.yak.ops.business.development.directory.DevelopmentDirectoryService;
 import io.yak.ops.business.development.domain.DevelopmentDirectory;
-import io.yak.ops.business.development.service.DevelopmentDirectoryService;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.web.bind.annotation.DeleteMapping;
