+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/controller/v1/DevelopmentEditorSettingsController.java
@@ -3,7 +3,7 @@ package io.yak.ops.business.development.controller.v1;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import io.yak.framework.common.Result;
-import io.yak.ops.business.development.service.DevelopmentEditorSettingsService;
+import io.yak.ops.business.development.editor.DevelopmentEditorSettingsService;
 import java.security.Principal;
 import java.util.Map;
 import org.springframework.web.bind.annotation.GetMapping;
