+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/domain/DevelopmentDraftConflictException.java
@@ -1,0 +1,9 @@
+package io.yak.ops.business.development.domain;
+
+/** Raised when an authoring command uses a stale optimistic draft revision. */
+public class DevelopmentDraftConflictException extends RuntimeException {
+
+  public DevelopmentDraftConflictException(String message) {
+    super(message);
+  }
+}
