+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/editor/DevelopmentEditorSettingsService.java
@@ -1,4 +1,4 @@
-package io.yak.ops.business.development.service;
+package io.yak.ops.business.development.editor;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
