+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/lineage/DevelopmentLineageOutbox.java
@@ -1,6 +1,5 @@
-package io.yak.ops.business.development.service;
+package io.yak.ops.business.development.lineage;
 
-import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.jdbc.core.JdbcTemplate;
