+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/lineage/DevelopmentLineageWorker.java
@@ -1,9 +1,10 @@
-package io.yak.ops.business.development.service;
+package io.yak.ops.business.development.lineage;
 
 import io.yak.ops.business.development.domain.DevelopmentNode;
 import io.yak.ops.business.development.domain.DevelopmentTaskRevision;
 import io.yak.ops.business.development.repository.DevelopmentNodeRepository;
 import io.yak.ops.business.development.repository.DevelopmentTaskRevisionRepository;
+import io.yak.ops.business.development.service.DevelopmentSqlLineageService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
