+++ yak-ops-business/yak-ops-business-data-development/src/main/java/io/yak/ops/business/development/lineage/DevelopmentLineageWriteTransaction.java
@@ -1,7 +1,8 @@
-package io.yak.ops.business.development.service;
+package io.yak.ops.business.development.lineage;
 
 import io.yak.ops.business.development.domain.DevelopmentNode;
 import io.yak.ops.business.development.domain.DevelopmentTaskRevision;
+import io.yak.ops.business.development.service.DevelopmentSqlLineageService;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
